添加最基本的缓存功能，每个类别缓存上次刷新的前十条目录。

+++ HFCampus/TopLineCell.m
@@ -39,7 +39,7 @@
 
 -(void)EScrollerViewDidClicked:(NSUInteger)index
 {
-    NSLog(@"index--%ld",index);
+    NSLog(@"index--%ld",(unsigned long)index);
 }
 
 - (void)setSelected:(BOOL)selected animated:(BOOL)animated
